Termine de hacer los lanzamientos de bombas. Hice mas pruebas en integracion, bombita, etc y cambie algunos detalles que nos faltaban del enunciado (por ejemplo, que el efecto del articulo dure hasta que se muera). Hice que Salida no se pueda destruir y saque el metodo ObtenerAdyacentes (no se usaba en ninguna parte, estaba al pedo)

+++  tp-algoritmos-tres/TP2-Bomberman/TP2-Bomberman/src/Interfaces/IDependienteDelTiempo.cs
@@ -8,6 +8,6 @@ namespace TP2_Bomberman.src
     // La implementan los elementos que dependen del tiempo para realizar una accion
     public interface IDependienteDelTiempo
     {
-        void CuandoPaseElTiempo(int tiempo);
+        void CuandoPaseElTiempo(double tiempo);
     }
 }
